implement matchers for different types

+++ math/math.go
@@ -2,6 +2,6 @@ package math
 
 import m "math"
 
-func AreEqual(x float32, y float32) bool {
+func AreEqual(x float64, y float64) bool {
 	return m.Abs(float64(x)-float64(y)) < 1e-6
 }
